fix: Court.Position type and duplicate log files
Fixed two issues discovered during MCP Inspector testing:

Issue 1: Court.Position Type Mismatch
- Changed Position from int? to decimal?
- CourtListener API returns decimal (e.g., 1.0) not integer
- Error: "JSON value could not be converted to System.Nullable`1[System.Int32]"
- Caused get_court tool to fail for all courts

Court.cs Changes:
- Line 19: public int? Position → public decimal? Position
- Added comment explaining API returns decimal

Issue 2: Duplicate Log Files
- Program.cs and appsettings.json both created file sinks
- Result: server20251013.log AND server-20251013.log
- Removed duplicate file sink from Program.cs

Program.cs Changes:
- Simplified Serilog configuration
- Now only uses appsettings.json configuration
- Removed: .Enrich.FromLogContext(), .WriteTo.Console(), .WriteTo.File()
- Kept: config.ReadFrom.Configuration() only

Log File Configuration (appsettings.json):
- Path: logs/server-.log (creates server-YYYYMMDD.log)
- Format: JSON
- Rolling: Daily
- Retention: 7 days
- Size limit: 1MB

Testing:
✅ Build successful
✅ get_court tool now works correctly
✅ Only one log file created

Discovered by: MCP Inspector testing with get_court tool

+++ CourtListener.MCP.Server/Models/Entities/Court.cs
@@ -16,5 +16,5 @@ public class Court
     public bool? InUse { get; set; }
     public bool? HasOpinionScraper { get; set; }
     public bool? HasOralArgumentScraper { get; set; }
-    public int? Position { get; set; }
+    public decimal? Position { get; set; }  // API returns decimal (e.g., 1.0) not integer
 }
